fix: abstraction for common stuff with reminders
