+++ raven/raven_messaging/doctype/raven_message/raven_message.py
@@ -966,11 +966,6 @@ class RavenMessage(Document):
 		# delete all the reactions for the message
 		frappe.db.delete("Raven Message Reaction", {"message": self.name})
 
-		# A dispatched scheduled message keeps a `sent_message` link to the message
-		# it produced — its record has no life of its own once the message is gone,
-		# and the link would otherwise block this delete.
-		frappe.db.delete("Raven Scheduled Message", {"sent_message": self.name})
-
 		# The deleted message may be sitting in clients' unread-notification badges:
 		# a mention of someone, or reactions on the owner's message. Those id sets
 		# are kept live by events — the rows vanishing from the DB doesn't reach the
@@ -994,6 +989,11 @@ class RavenMessage(Document):
 				user=self.owner,
 				after_commit=True,
 			)
+
+		# A dispatched scheduled message keeps a `sent_message` link to the message
+		# it produced — its record has no life of its own once the message is gone,
+		# and the link would otherwise block this delete.
+		frappe.db.delete("Raven Scheduled Message", {"sent_message": self.name})
 
 		# If this was the channel's latest message, recompute the channel summary to the
 		# PREVIOUS message instead of just nulling it — otherwise the sidebar teaser keeps
